feat: Add optional dropout layer to neural network construction
The changes made in this commit add an optional dropout layer at the end of the neural network construction in the `create_model` function in `main.py`. A new parameter `dropout_prob` has been added to the function, which controls whether the dropout layer is used and its probability.

The `parse_args` function has also been updated to include a new command-line argument `--dropout` to set the dropout probability. If the dropout probability is set to 0 (which is the default), no dropout layer will be added.

Finally, the `main` function has been updated to use the new `dropout_prob` parameter when creating the model.

These changes will allow users to easily add a dropout layer to the neural network to help prevent overfitting.